requirements.txt
psutil
pynvml
pyserial
prometheus_client
pyarrow
numpy
tomli; python_version < '3.11'
